Move noise generation machinery to seperate 'photometric_noise' submodule
