+++ src/synthesizer/imaging/image_generators.py
@@ -1673,320 +1673,3 @@ def _prepare_galaxy_image_labels(
                 component_labels_by_emitter[emitter].append(label)
 
     return galaxy_combine_labels, component_labels_by_emitter
-
-
-def _cf_periodicity_dilution_correction_standalone(
-    cf_shape,
-) -> np.ndarray:
-    """Calculates the correction factor for DFT-based Correlation Functions.
-
-    to account for the assumption of periodicity.
-    Ported from an internal GalSim function.
-
-    Args:
-        cf_shape: Tuple (Ny, Nx) representing the shape of the
-        correlation function array.
-
-    Returns:
-        A 2D NumPy array with the correction factors.
-    """
-    ny, nx = cf_shape
-
-    # Create coordinate arrays for frequency domain
-    # dx_coords corresponds to frequencies in the x-direction (columns)
-    # dy_coords corresponds to frequencies in the y-direction (rows)
-    dx_coords = np.fft.fftfreq(nx) * float(nx)
-    dy_coords = np.fft.fftfreq(ny) * float(ny)
-
-    # Create 2D grids of these coordinates
-    # deltax will have shape (Ny, Nx) and vary along columns (axis 1)
-    # deltay will have shape (Ny, Nx) and vary along rows (axis 0)
-    deltax, deltay = np.meshgrid(dx_coords, dy_coords)
-
-    denominator = (nx - np.abs(deltax)) * (ny - np.abs(deltay))
-
-    # Avoid division by zero if denominator entries are zero,
-    # though this is unlikely with standard fftfreq outputs for valid shapes.
-    # np.finfo(float).eps could be added for numerical stability if needed.
-    valid_denominator = np.where(
-        denominator == 0, 1.0, denominator
-    )  # Avoid true div by zero
-
-    correction = (float(nx * ny)) / valid_denominator
-    if np.any(
-        denominator == 0
-    ):  # Put back zeros where they were to avoid infs if original was 0
-        correction[denominator == 0] = 0  # Or handle as error/warning
-
-    return correction
-
-
-def _generate_noise_from_rootps_standalone(
-    rng: np.random.Generator, shape, rootps: np.ndarray
-) -> np.ndarray:
-    """Generates a real-space noise field from its sqrt(PowerSpectrum).
-
-    Ported and adapted from an internal GalSim function.
-
-    Args:
-        rng: NumPy random number generator.
-        shape: Tuple (Ny, Nx) of the output real-space noise field.
-        rootps: The half-complex 2D array (Ny, Nx//2 + 1) representing
-                the square root of the Power Spectrum from rfft2.
-
-    Returns:
-        A 2D NumPy array representing the generated correlated noise field.
-    """
-    ny, nx = shape
-
-    # GalSim's GaussianDeviate implies specific scaling for random numbers.
-    # E[|gvec_k|^2] = Ny * Nx for each complex component k
-    # if parts are N(0, 0.5*Ny*Nx).
-    sigma_val_for_gvec_parts = np.sqrt(0.5 * ny * nx)
-
-    gvec_real = rng.normal(scale=sigma_val_for_gvec_parts, size=rootps.shape)
-    gvec_imag = rng.normal(scale=sigma_val_for_gvec_parts, size=rootps.shape)
-    gvec = gvec_real + 1j * gvec_imag
-
-    # Impose Hermitian symmetry properties and scaling for DC/Nyquist terms.
-    # This ensures the iFFT results in a real field
-    # with correct variance distribution.
-    rt2 = np.sqrt(2.0)
-
-    # DC component (ky=0, kx=0)
-    gvec[0, 0] = rt2 * gvec[0, 0].real
-
-    # Nyquist frequency for y-axis (ky=Ny/2) at kx=0 (if Ny is even)
-    if ny % 2 == 0:
-        gvec[ny // 2, 0] = rt2 * gvec[ny // 2, 0].real
-
-    # Nyquist frequency for x-axis (kx=Nx/2) at ky=0 (if Nx is even)
-    if nx % 2 == 0:
-        gvec[0, nx // 2] = rt2 * gvec[0, nx // 2].real
-
-    # Corner Nyquist component (ky=Ny/2, kx=Nx/2) (if both Ny and Nx are even)
-    if ny % 2 == 0 and nx % 2 == 0:
-        gvec[ny // 2, nx // 2] = rt2 * gvec[ny // 2, nx // 2].real
-
-    # Conjugate symmetry for the kx=0 column (y-axis frequencies)
-    # gvec[Ny-row, 0] = conj(gvec[row, 0]) for row in 1 .. (Ny/2 - 1)
-    # or ((Ny-1)/2)
-    # This handles the negative y-frequencies for kx=0.
-    # The slice `ny-1 : ny//2 : -1` covers rows from Ny-1 down to (Ny//2 + 1).
-    # The slice `1 : (ny+1)//2` covers rows from 1 up to Ny//2.
-    if ny > 1:  # This symmetry operation is relevant
-        # if Ny > 1 (or Ny > 2 for some ranges)
-        gvec[ny - 1 : ny // 2 : -1, 0] = np.conj(gvec[1 : (ny + 1) // 2, 0])
-
-    # Conjugate symmetry for the kx=Nx/2 column (if Nx is even)
-    # This handles negative y-frequencies for the Nyquist x-frequency.
-    if nx % 2 == 0:
-        kx_nyq_idx = nx // 2
-        if ny > 1:
-            gvec[ny - 1 : ny // 2 : -1, kx_nyq_idx] = np.conj(
-                gvec[1 : (ny + 1) // 2, kx_nyq_idx]
-            )
-
-    # Element-wise multiplication in Fourier space
-    noise_field_k_space = gvec * rootps
-
-    # Inverse FFT to get the real-space noise field
-    # The `s` parameter ensures the output shape matches the desired `shape`.
-    noise_real_space = np.fft.irfft2(noise_field_k_space, s=shape)
-
-    return noise_real_space
-
-
-def _estimate_correlated_noise_cf(
-    source_arr: np.ndarray,
-    subtract_mean: bool = False,
-    correct_periodicity: bool = True,
-) -> np.ndarray:
-    """Estimate the correlation function (CF) of a 2D noise array.
-
-    Computes an unrolled correlation function (DFT convention, origin at
-    [0, 0]) from the power spectrum of the source array. The result is
-    intended to be cached and later consumed by ``_generate_correlated_noise``
-    to produce noise realisations for images of arbitrary shape.
-
-    This function is deliberately separated from noise generation so that the
-    expensive FFT step can be computed once (e.g. per instrument filter) and
-    reused across many images.
-
-    Args:
-        source_arr (np.ndarray):
-            A 2D float array of observed noise (e.g. a blank-sky cutout).
-            Must already have units stripped before calling.
-        subtract_mean (bool):
-            If True the DC component of the power spectrum is zeroed, which
-            effectively removes the mean offset from the noise model.
-        correct_periodicity (bool):
-            If True a correction factor is applied to compensate for the
-            assumption of periodicity in the DFT.
-
-    Returns:
-        np.ndarray:
-            A 2D array of the same shape as ``source_arr`` holding the
-            unrolled CF with the origin at [0, 0].
-    """
-    if source_arr.ndim != 2:
-        raise ValueError("source_arr must be a 2D numpy array.")
-
-    source_shape = source_arr.shape
-
-    ft_array = np.fft.rfft2(source_arr)
-    ps_array = np.abs(ft_array) ** 2
-    ps_array /= np.prod(source_shape)
-
-    if subtract_mean:
-        ps_array[0, 0] = 0.0
-
-    cf_array = np.fft.irfft2(ps_array, s=source_shape)
-
-    if correct_periodicity:
-        correction = _cf_periodicity_dilution_correction_standalone(
-            source_shape
-        )
-        cf_array *= correction
-
-    return cf_array
-
-
-def _generate_correlated_noise(
-    cf_source: np.ndarray,
-    target_shape: tuple,
-    rng_seed: int = None,
-) -> np.ndarray:
-    """Generate a correlated noise realisation for a given target shape.
-
-    Uses a pre-computed correlation function (CF) — typically returned by
-    ``_estimate_correlated_noise_cf`` and cached at the call site — to
-    produce a noise field whose spatial statistics match those of the
-    original noise template.  The CF is resampled to the target pixel
-    dimensions before noise generation.
-
-    Args:
-        cf_source (np.ndarray):
-            The unrolled 2D correlation function (DFT convention, origin at
-            [0, 0]) previously returned by ``_estimate_correlated_noise_cf``.
-        target_shape (tuple of int):
-            The ``(ny, nx)`` pixel dimensions of the desired noise field.
-        rng_seed (int, optional):
-            Seed for the random number generator.  Pass the same seed to
-            reproduce an identical noise realisation.
-
-    Returns:
-        np.ndarray:
-            A 2D float array of shape ``target_shape`` containing the
-            generated correlated noise field.
-
-    Raises:
-        InconsistentArguments:
-            If the source CF is too small to contribute any pixels to the
-            target CF grid.
-    """
-    rng = np.random.default_rng(rng_seed)
-    source_shape = cf_source.shape
-
-    if source_shape[0] < target_shape[0] or source_shape[1] < target_shape[1]:
-        warn(
-            "Source noise map is smaller than the target image, which may"
-            " result in truncation of the noise model."
-        )
-
-    # Roll source CF so that the peak (variance, at [0,0]) moves to the centre
-    cf_source_rolled = np.roll(
-        cf_source,
-        shift=(source_shape[0] // 2, source_shape[1] // 2),
-        axis=(0, 1),
-    )
-
-    # Copy the central region of the source CF into an array of target_shape
-    cf_target_rolled = np.zeros(target_shape, dtype=float)
-
-    src_cy, src_cx = source_shape[0] // 2, source_shape[1] // 2
-    trg_cy, trg_cx = target_shape[0] // 2, target_shape[1] // 2
-
-    y_start_src = max(0, src_cy - trg_cy)
-    y_end_src = min(source_shape[0], src_cy + (target_shape[0] - trg_cy))
-    x_start_src = max(0, src_cx - trg_cx)
-    x_end_src = min(source_shape[1], src_cx + (target_shape[1] - trg_cx))
-
-    y_start_trg = max(0, trg_cy - src_cy)
-    y_end_trg = min(target_shape[0], trg_cy + (source_shape[0] - src_cy))
-    x_start_trg = max(0, trg_cx - src_cx)
-    x_end_trg = min(target_shape[1], trg_cx + (source_shape[1] - src_cx))
-
-    dy = min(y_end_src - y_start_src, y_end_trg - y_start_trg)
-    dx = min(x_end_src - x_start_src, x_end_trg - x_start_trg)
-
-    if dy <= 0 or dx <= 0:
-        raise exceptions.InconsistentArguments(
-            "Source CF is too small to contribute to the target CF. "
-            "Ensure the source noise map is large enough to model noise "
-            "for the target image."
-        )
-
-    cf_target_rolled[
-        y_start_trg : y_start_trg + dy, x_start_trg : x_start_trg + dx
-    ] = cf_source_rolled[
-        y_start_src : y_start_src + dy, x_start_src : x_start_src + dx
-    ]
-
-    # Unroll back to DFT convention (origin at [0, 0])
-    cf_on_target_grid_unrolled = np.roll(
-        cf_target_rolled,
-        shift=(-(target_shape[0] // 2), -(target_shape[1] // 2)),
-        axis=(0, 1),
-    )
-
-    ps_target_fft = np.fft.rfft2(cf_on_target_grid_unrolled)
-    rootps_target = np.sqrt(np.abs(ps_target_fft))
-
-    return _generate_noise_from_rootps_standalone(
-        rng, target_shape, rootps_target
-    )
-
-
-def _model_and_apply_correlated_noise(
-    source_image_arr: np.ndarray,
-    target_image_arr: np.ndarray,
-    subtract_mean: bool = False,
-    correct_periodicity: bool = True,
-    rng_seed: int = None,
-) -> np.ndarray:
-    """Model correlated noise from a source image and apply it to a target.
-
-    This is a convenience wrapper around ``_estimate_correlated_noise_cf``
-    and ``_generate_correlated_noise``.  Prefer calling those two functions
-    directly when the correlation function should be cached across multiple
-    calls (e.g. when applying the same noise model to many images).
-
-    Args:
-        source_image_arr (np.ndarray):
-            A 2D array used to model the correlated noise characteristics.
-        target_image_arr (np.ndarray):
-            A 2D array to which the generated correlated noise is added.
-        subtract_mean (bool):
-            If True the DC component of the power spectrum is zeroed before
-            estimating the CF. Default is False.
-        correct_periodicity (bool):
-            If True a correction factor is applied to compensate for the
-            assumption of periodicity in the DFT. Default is True.
-        rng_seed (int, optional):
-            Seed for the random number generator for reproducibility.
-
-    Returns:
-        np.ndarray:
-            ``target_image_arr`` with the synthesised correlated noise added.
-    """
-    if source_image_arr.ndim != 2 or target_image_arr.ndim != 2:
-        raise ValueError("Input images must be 2D numpy arrays.")
-
-    cf = _estimate_correlated_noise_cf(
-        source_image_arr,
-        subtract_mean=subtract_mean,
-        correct_periodicity=correct_periodicity,
-    )
-    noise = _generate_correlated_noise(cf, target_image_arr.shape, rng_seed)
-    return target_image_arr + noise
